gui: wizard: moved components in a dedicated folder
