+++ circleguard/wizard/pages/apikey.py
@@ -1,5 +1,5 @@
-from wizard.page import WizardPage
-from wizard.label import WizardLabel
+from wizard.components.page import WizardPage
+from wizard.components.label import WizardLabel
 
 from widgets import LineEditSetting
 from PyQt6.QtWidgets import QGridLayout
